+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [497] - 2026-03-19
+
+- [TASK-703] Add lint rule: issue tasks must have at least one test-type criterion
+
 ## [496] - 2026-03-19
 
 - [TASK-699] Add domain_test_commands to KNOWN_KEYS in tusk-lint rule 7